tests/wpt/tests/css/css-conditional/container-queries/scroll-state/at-container-scroll-direction-parsing.html

<!DOCTYPE html>
<title>CSS Conditional Test: @container scroll-state(direction) query parsing</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#container-rule">
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#direction">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<div style="container-name:name;container-type:scroll-state">
  <main id="cq-main"></main>
</div>
<script>
  setup(() => assert_implements_scroll_state_container_queries());

  test_cq_condition_known('scroll-state(direction)');
  test_cq_condition_known('scroll-state(direction: none)');
  test_cq_condition_known('scroll-state(direction: top)');
  test_cq_condition_known('scroll-state(direction: left)');
  test_cq_condition_known('scroll-state(direction: bottom)');
  test_cq_condition_known('scroll-state(direction: right)');
  test_cq_condition_known('scroll-state(direction: inline-start)');
  test_cq_condition_known('scroll-state(direction: inline-end)');
  test_cq_condition_known('scroll-state(direction: block-start)');
  test_cq_condition_known('scroll-state(direction: block-end)');
  test_cq_condition_known('scroll-state(direction: block)');
  test_cq_condition_known('scroll-state(direction: x)');
  test_cq_condition_known('scroll-state(direction: y)');
  test_cq_condition_known('scroll-state(direction: inline)');
  test_cq_condition_known('(scroll-state(direction: inline-end))');
  test_cq_condition_known('scroll-state((direction: left))');
  test_cq_condition_known('scroll-state(not ((direction: bottom) and (direction: right)))');
  test_cq_condition_known('scroll-state((direction: left) or (direction: top))');

  test_cq_condition_unknown('scroll-state(direction: auto)');
  test_cq_condition_unknown('scroll-state(direction: true)');
  test_cq_condition_unknown('scroll-state(style(direction: left))');
  test_cq_condition_unknown('style(scroll-state(direction: left))');
  test_cq_condition_unknown('scroll-state(direction:)');
</script>
